Backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.1
opencv-python==4.9.0.80
